refractored algorithim & changed score data-type from int to double (#76)
* refractored algorithim & changed score data-type from int to double

* renamed variables and documentation to explain process accurately

+++ Project_2/LocationStatusAPI/src/main/java/com/Project_2_Location_Status_API/Entities/Status.java
@@ -21,7 +21,7 @@ public class Status {
     private int id;
 
     @Column(name = "score")
-    private Integer score;
+    private double score;
 
     @Column(name = "location")
     private String location;
